1.4.0 - Added Mouse Tracking Code
Added a script to the HTML to write the mouse x and y co-ordinates to the divShow division.
To support this, I've added code to the createTime.js file to get the mouse x and y co-ords.
I left in the writing to the canvas code in case we need to use it. Might change the js files name later to be more descriptive.