v1.6.2 Fixed Recursion Issue
Changed the mouse listeners and removed the recursion which has fixed the unwanted firefox breaking. However this lead to weird behaviour when clicking the black square. After the first time the black square is clicked the emulateVibrate() function runs and then the mouse listeners stop responding. I commented out the call to emulate vibrate which stops this happening and everything is fine now; since we don't need that exact functionality for the real emulator we may not have to worry why the freaking hell emulateVibrate() ruins things when not inside an evil recursive method.