1.13.1 - Cleaned Bits up
That new date object bit worked excellently.
You pass it a year and a month and it creates a date object of that month in that year
and returns it at the first day of that month in that year as an integer. It's a shortcut so 
we don't actually have to calculate anything for each month, we can just use that to set up
each month as you go along....