1.17.3 - New idea for reminders
It now prints the date to the white square (which I might make another canvas at some point soon)
and I think I'll add little clickable next and back buttons to change the date that is printed.
This will allow people to change the day to set a reminder on!