1.18.2 - Saving context doesn't work. So using Javascript to return.
This is because the save and restore work on things drawn to the canvas and
I don't think the Javascript functions we have count, technically.
Either way, reinitialising the prototype works, so I'll work on building something to 'save'
the month the person was on last so they can return to that.